Encapsulate config properties in accessors Default mutations is 0.2 (100/20000 x 40) Set historic and modern kill constant to 1 and revert modern birth rate to same as historic birth rate Change behaviour of capacity tuning to produce a smooth transition between epochs Record average capacity in generations result

+++ src/main/java/uk/edu/populationfitness/output/TuningReader.java
@@ -30,7 +30,7 @@ public class TuningReader {
         tuning.min_fitness = Double.parseDouble(row[9]);
         tuning.max_fitness = Double.parseDouble(row[10]);
         if (row.length > 11){
-            tuning.mutations_per_gene = Integer.parseInt(row[11]);
+            tuning.mutations_per_gene = Double.parseDouble(row[11]);
         }
         if (row.length > 12){
             tuning.series_runs = Integer.parseInt(row[12]);
